Rename constructor slice
