+++ package.json
@@ -7,8 +7,6 @@
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
-    "@types/jest": "^27.5.2",
-    "@types/node": "^16.18.23",
     "@types/react": "^18.0.31",
     "@types/react-dom": "^18.0.11",
     "@types/uuid": "^9.0.1",
@@ -43,8 +41,8 @@
     "@storybook/test": "^7.6.10",
     "@testing-library/react": "^14.1.2",
     "@testing-library/user-event": "^14.5.1",
-    "@types/jest": "^29.5.11",
-    "@types/node": "^20.10.5",
+    "@types/jest": "^29.5.14",
+    "@types/node": "^20.19.1",
     "@types/react": "^18.2.45",
     "@types/react-dom": "^18.2.18",
     "@types/react-test-renderer": "^18.0.7",
