+++ src/components/burger-ingredient/burger-ingredient.tsx
@@ -1,7 +1,7 @@
 import { FC, memo } from 'react';
 import { useLocation } from 'react-router-dom';
 import { useDispatch } from '../../services/store';
-import { addIngredient } from '../../services/constructor/slice';
+import { addIngredient } from '../../services/burger-constructor/slice';
 
 import { BurgerIngredientUI } from '@ui';
 import { TBurgerIngredientProps } from './type';
